% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_code.R
\name{backticks_for_inline_code}
\alias{backticks_for_inline_code}
\title{Foobar}
\usage{
backticks_for_inline_code(x)
}
\arguments{
\item{x}{Same in param: \code{"hi there"}}
}
\description{
Quotation marks in inline code: \code{"hi there"}.
Plays nice with equation: \eqn{\partial Y / \partial X = a + \varepsilon/2}
}
