% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesIf_true.R
\name{examplesIf_true}
\alias{examplesIf_true}
\title{Examples If TRUE}
\usage{
examplesIf_true()
}
\arguments{
\item{x}{A parameter}
}
\value{
Some value
}
\description{
Examples If TRUE
}
\examples{
\dontshow{if (2 + 2 == 4) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
examplesIf_true()
\dontshow{\}) # examplesIf}
}
