% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_dontrun.R
\name{hello_dontrun}
\alias{hello_dontrun}
\title{Mixed runnable and non-runnable examples}
\usage{
hello_dontrun(x = 2)
}
\arguments{
\item{x}{A parameter}
}
\value{
Some value
}
\description{
Mixed runnable and non-runnable examples
}
\examples{
hello_dontrun()

\dontrun{
stop("this must never be evaluated")
}

hello_dontrun()

\donttest{
Sys.sleep(1000)
}
}
